+++ web/apps/photos/src/services/wasmHeicConverter/wasmHEICConverterService.ts
@@ -1,11 +1,11 @@
 import { CustomError } from "@ente/shared/error";
 import { addLogLine } from "@ente/shared/logging";
+import { retryAsyncFunction } from "@ente/shared/promise";
 import { logError } from "@ente/shared/sentry";
 import { convertBytesToHumanReadable } from "@ente/shared/utils/size";
 import { ComlinkWorker } from "@ente/shared/worker/comlinkWorker";
 import QueueProcessor from "services/queueProcessor";
 import { getDedicatedConvertWorker } from "utils/comlink/ComlinkConvertWorker";
-import { retryAsyncFunction } from "utils/network";
 import { DedicatedConvertWorker } from "worker/convert.worker";
 
 const WORKER_POOL_SIZE = 2;
